requirements.txt
browser-use-sdk
langchain-aws
openai
python-dotenv
pydantic>=2.0
boto3
